chore: organize search packages files
