+++ lib/src/core/controller/states/search_packages_states.dart
@@ -1,0 +1,15 @@
+import '../../data/models/packages_model.dart';
+
+abstract interface class SearchPackagesState {}
+
+final class IdleSearchPackagesState extends SearchPackagesState {}
+
+final class LoadingSearchPackagesState extends SearchPackagesState {}
+
+final class LoadedSearchPackagesState extends SearchPackagesState {
+  final PackagesModel packagesModel;
+
+  LoadedSearchPackagesState({required this.packagesModel});
+}
+
+final class ErrorSearchPackagesState extends SearchPackagesState {}
