+++ lib/src/core/widgets/modal_search_packages/erros_searchs_packages/error_search_packages_informations.dart
