+++ lib/src/core/widgets/modal_search_packages/loading_search_packages.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
-import '../../../../core/themes/app_images.dart';
-import '../../../../core/widgets/lotties_widgets/lottie_assets.dart';
+import '../../themes/app_images.dart';
+import '../lotties_widgets/lottie_assets.dart';
 
 class LoadingSearchPackages extends StatelessWidget {
   const LoadingSearchPackages({
